+++ addons/confirma/src/classes/TestingClass.cs
@@ -5,6 +5,7 @@ using System.Reflection;
 using Confirma.Attributes;
 using Confirma.Helpers;
 using Confirma.Types;
+using Godot;
 
 namespace Confirma.Classes;
 
@@ -40,8 +41,10 @@ public class TestingClass
 
             if (!TestMethods.Any())
             {
-                Log.PrintError($"No test Methods found with the name '{props.MethodName}'.");
-                return new TestClassResult(0, 0, 0, 1);
+                Log.PrintError(
+                    $"No test Methods found with the name '{props.MethodName}'."
+                );
+                return new(0, 0, 0, 1);
             }
         }
 
@@ -49,9 +52,20 @@ public class TestingClass
         {
             warnings += RunLifecycleMethod("SetUp");
 
+            int currentOrphans = GetOrphans();
+
             TestMethodResult methodResult = method.Run(props);
 
             warnings += RunLifecycleMethod("TearDown");
+
+            int newOrphans = GetOrphans();
+            if (currentOrphans < newOrphans)
+            {
+                warnings++;
+                Log.PrintWarning(
+                    $"Calling {method.Name} created {newOrphans - currentOrphans} new orphan/s.\n"
+                );
+            }
 
             passed += methodResult.TestsPassed;
             failed += methodResult.TestsFailed;
@@ -109,5 +123,14 @@ public class TestingClass
         }
 
         return method.HasMultiple ? (byte)1 : (byte)0;
+    }
+
+    private int GetOrphans()
+    {
+        return !_props.MonitorOrphans
+            ? 0
+            : (int)Performance.Singleton.GetMonitor(
+                Performance.Monitor.ObjectOrphanNodeCount
+            );
     }
 }