+++ addons/confirma/src/types/TestResult.cs
@@ -6,6 +6,7 @@ public record TestResult
     public uint TestsPassed { get; set; }
     public uint TestsFailed { get; set; }
     public uint TestsIgnored { get; set; }
+    public uint TotalOrphans { get; set; }
     public double TotalTime { get; set; }
     public uint Warnings { get; set; }
 
@@ -27,6 +28,7 @@ public record TestResult
         a.TotalTests += b.TestsPassed + b.TestsFailed + b.TestsIgnored;
         a.TestsPassed += b.TestsPassed;
         a.TestsFailed += b.TestsFailed;
+        a.TotalOrphans += b.TotalOrphans;
         a.TestsIgnored += b.TestsIgnored;
         a.Warnings += b.Warnings;